完善用户模块与项目模块逻辑

+++ src/main/java/com/program/itta/service/impl/ItemServiceImpl.java
@@ -1,6 +1,9 @@
 package com.program.itta.service.impl;
 
+import com.program.itta.common.config.JwtConfig;
 import com.program.itta.common.exception.GlobalExceptionHandler;
+import com.program.itta.common.exception.item.ItemNameExistsException;
+import com.program.itta.common.exception.item.ItemNotExistsException;
 import com.program.itta.domain.entity.Item;
 import com.program.itta.mapper.ItemMapper;
 import com.program.itta.service.ItemService;
@@ -9,6 +12,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import javax.annotation.Resource;
 import java.util.Date;
 import java.util.List;
 
@@ -26,6 +30,10 @@ public class ItemServiceImpl implements ItemService {
     @Autowired
     private ItemMapper itemMapper;
 
+
+    @Resource
+    private JwtConfig jwtConfig;
+
     /**
      * 添加项目
      *
@@ -34,6 +42,12 @@ public class ItemServiceImpl implements ItemService {
      */
     @Override
     public Boolean addItem(Item item) {
+        Integer userId = jwtConfig.getUserId();
+        item.setLeaderId(userId);
+        Boolean judgeItem = judgeItemName(item);
+        if (judgeItem) {
+            throw new ItemNameExistsException("项目名称已存在");
+        }
         int insert = itemMapper.insert(item);
         if (insert != 0) {
             logger.info("添加项目：" + item.getName());
@@ -50,6 +64,10 @@ public class ItemServiceImpl implements ItemService {
      */
     @Override
     public Boolean deleteItem(Item item) {
+        Boolean judgeItemExists = judgeItemExists(item);
+        if (!judgeItemExists) {
+            throw new ItemNotExistsException("该项目不存在，项目id查找为空");
+        }
         int delete = itemMapper.deleteByPrimaryKey(item.getId());
         if (delete != 0) {
             logger.info("删除项目：" + item.getName() + "项目id为:" + item.getId());
@@ -66,6 +84,10 @@ public class ItemServiceImpl implements ItemService {
      */
     @Override
     public Boolean updateItem(Item item) {
+        Boolean judgeItemExists = judgeItemExists(item);
+        if (!judgeItemExists) {
+            throw new ItemNotExistsException("该项目不存在，项目id查找为空");
+        }
         item.setUpdateTime(new Date());
         int update = itemMapper.updateByPrimaryKey(item);
         if (update != 0) {
@@ -76,23 +98,30 @@ public class ItemServiceImpl implements ItemService {
     }
 
     @Override
-    public Boolean judgeItem(Item item) {
+    public List<Item> selectAllItem(List<Integer> itemIdList) {
+        List<Item> itemList = null;
+        for (int i = 0; i < itemIdList.size(); i++) {
+            Item item = itemMapper.selectByPrimaryKey(itemIdList.get(i));
+            itemList.add(item);
+        }
+        return itemList;
+    }
+
+    private Boolean judgeItemName(Item item) {
         List<Item> items = itemMapper.selectAllItemByUserId(item.getLeaderId());
         for (Item item1 : items) {
-            if (item1.getName().equals(item.getName())){
+            if (item1.getName().equals(item.getName())) {
                 return true;
             }
         }
         return false;
     }
 
-    @Override
-    public List<Item> selectAllItem(List<Integer> itemIdList) {
-        List<Item> itemList = null;
-        for (int i = 0; i < itemIdList.size(); i++){
-            Item item = itemMapper.selectByPrimaryKey(itemIdList.get(i));
-            itemList.add(item);
+    private Boolean judgeItemExists(Item item) {
+        Item select = itemMapper.selectByPrimaryKey(item.getId());
+        if (select != null) {
+            return true;
         }
-        return itemList;
+        return false;
     }
 }
